moved App.js and updated index import, added folders for structure and base files for some sections, created initial Styles folder and global styles files, added a couplee for filler
