+++ React/src/App.js
@@ -1,0 +1,16 @@
+// components
+import PageHeader from './components/PageHeader'
+import Home from './pages/Home'
+import PageFooter from './components/PageFooter'
+
+const App = () => {
+  return (
+    <>
+      <PageHeader />
+      <Home />
+      <PageFooter />
+    </>
+  )
+}
+
+export default App
