--- React/src/components/App.js
@@ -1,13 +1,0 @@
-// components
-// TODO: import component here, use this format ' import ComponentName from './FileName' '
-import PageHeader from "./PageHeader"
-
-const App = () => {
-  return (
-    <>
-      <PageHeader />
-    </>
-  )
-}
-
-export default App
