+++ React/src/components/CategoryWrapper.js
@@ -1,0 +1,24 @@
+// styles
+import '../styles/GlobalStyles.css'
+
+const CategoryWrapper = () => {
+    // const categories = [
+    //     'Gen 1',
+    //     'Gen 2',
+    //     'Gen 3',
+    //     'Gen 4',
+    //     'Gen 5',
+    //     'Gen 6',
+    //     'Gen 7',
+    //     'Gen 8',
+    //     'Gen 9',
+    //     'Custom Games'
+    // ]
+    return (
+        <>
+            <h3 class='section-header'>Select a Category</h3>
+        </>
+    )
+}
+
+export default CategoryWrapper
