+++ React/src/components/HomePageHeader.js
@@ -1,0 +1,10 @@
+const HomePageHeader = () => {
+    return (
+        <>
+            <div>Nuzlock Tracker App</div>
+            <div>Created by: 2tummies and Urason</div>
+        </>
+    )
+}
+
+export default HomePageHeader
