+++ React/src/components/PageFooter.js
@@ -1,0 +1,8 @@
+const PageFooter = () => {
+    return (
+        <>
+        </>
+    )
+}
+
+export default PageFooter
