+++ React/src/components/PageHeader.js
@@ -1,8 +1,6 @@
 const PageHeader = () => {
     return (
         <>
-            <div>Nuzlock Tracker App</div>
-            <div>Created by: 2tummies and Urason</div>
         </>
     )
 }
