+++ React/src/styles/GlobalStyles.css
@@ -1,0 +1,7 @@
+.page-container {
+    padding: 1rem
+}
+
+.section-header {
+    text-align: center;
+}
